+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to the Pixel Banner plugin will be documented in this file.
 
+### [v2.20.3] - 2025-01-18
+#### ✨ Added
+- added `font weight` option to the banner icon settings (general, folder images, and frontmatter)
+
+#### 🐛 Fixed
+- fixed issue with a large gap appearing between the banner and the note content when using the `hide embedded note banners` setting
+
 ### [v2.20.2] - 2025-01-17
 #### ✨ Added
 - added `banner-fade-in-animation-duration` general setting to control the duration of the fade in animation for the banner image